# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_ExportData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[ExportData]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__object__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[dict[str, object]]",
        "type": "object"
      },
      "ApiStatus": {
        "enum": [
          "success",
          "accepted",
          "error"
        ],
        "title": "ApiStatus",
        "type": "string"
      },
...
      "ExportData": {
        "additionalProperties": false,
        "properties": {
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "download_url",
          "sha256",
          "filename"
        ],
        "title": "ExportData",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Traceable SEM nanoparticle analysis and evidence-backed query API",
    "title": "NanoLoop Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analyses/{job_id}/export": {
      "get": {
        "operationId": "exportAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExportData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Export Analysis",
        "tags": [
          "exports"
        ]
      }
    },
...
  }
}